backend_py/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
asyncpg==0.31.0
aiomqtt==2.3.0
python-socketio==5.11.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic-settings==2.3.0
httpx==0.27.0
qrcode[pil]==8.0
nanoid==2.0.0
slowapi==0.1.9
fastapi-mail==1.4.1
python-multipart==0.0.9
